requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.9
Pillow>=10.0.0
